Changed small things in the notation and phrases on sections 1.1 and 1.2.
I improved the text a little bit and also made small changes in the notation on those sections. If you think some changes didn't help at all, feel free to edit them out. For the moment, we can try to improve the notebook in english and then, later on, translate everything in one sitting. If we know we have everything correct, translating should be the easiest part. Overall, the notebook is pretty awesome and definetly would be a good way to review stuff for the test (after all the other topics are added).